+++ packages/plugin-sound-cloud/web/src/toolbar/soundCloudURLInputModal.jsx
@@ -43,7 +43,7 @@ export default class SoundCloudURLInputModal extends Component {
       <UrlInputModal
         onConfirm={this.onConfirm}
         helpers={helpers}
-        url={url}
+        input={url}
         t={t}
         languageDir={languageDir}
         title={
@@ -53,7 +53,7 @@ export default class SoundCloudURLInputModal extends Component {
         dataHook={'soundCloudUploadModal'}
         saveLabel={doneLabel}
         cancelLabel={cancelLabel}
-        setUrl={url => this.setState({ url })}
+        onInputChange={url => this.setState({ url })}
         errorMessage={t('SoundCloudUploadModal_Input_InvalidUrl')}
         placeholder={t('SoundCloudUploadModal_Input_Placeholder')}
         onCloseRequested={helpers.closeModal}