+++ packages/plugin-vertical-embed/web/src/createVerticalEmbedPlugin.js
@@ -1,0 +1,30 @@
+import { createBasePlugin } from 'wix-rich-content-editor-common';
+
+import { VERTICAL_EMBED_TYPE, DEFAULTS } from './constants';
+import VerticalEmbedComponent from './components/vertical-embed-component';
+import createToolbar from './toolbar';
+
+const createVerticalEmbedPlugin = (config = {}) => {
+  const type = VERTICAL_EMBED_TYPE;
+  const { helpers, theme, t, [type]: settings = {}, isMobile, ...rest } = config;
+
+  return createBasePlugin({
+    component: VerticalEmbedComponent,
+    settings,
+    theme,
+    type,
+    toolbar: createToolbar({
+      settings,
+      helpers,
+      t,
+      isMobile,
+    }),
+    helpers,
+    t,
+    defaultPluginData: DEFAULTS,
+    isMobile,
+    ...rest,
+  });
+};
+
+export { createVerticalEmbedPlugin, VERTICAL_EMBED_TYPE };