+++ packages/plugin-vertical-embed/web/src/index.js
@@ -1,0 +1,4 @@
+export { createVerticalEmbedPlugin } from './createVerticalEmbedPlugin';
+export { VERTICAL_EMBED_TYPE, verticalEmbedProviders } from './constants';
+export { Modals, ModalsMap } from './modals';
+export { pluginVerticalEmbed } from './editor';